phase-2-web/backend/src/config.py
"""
Configuration settings for FastAPI backend.

This module loads environment variables and provides a Settings object
for the application.
"""

import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()


class Settings:
    """
    Application settings loaded from environment variables.

    Attributes:
        database_url: PostgreSQL connection string (required)
        better_auth_url: Frontend URL for Better Auth JWKS endpoint
        frontend_url: Frontend URL for CORS
        debug: Enable debug mode (default: False)
        environment: Environment name (default: "development")
        r2_account_id: Cloudflare R2 account ID
        r2_access_key_id: R2 access key ID
        r2_secret_access_key: R2 secret access key
        r2_bucket_name: R2 bucket name
        r2_public_url: R2 public URL for accessing objects
    """

    def __init__(self):
        self.database_url = os.getenv("DATABASE_URL", "")
        self.better_auth_url = os.getenv("BETTER_AUTH_URL", "http://localhost:3000")
        self.frontend_url = os.getenv("FRONTEND_URL", "http://localhost:3000")
        self.debug = os.getenv("DEBUG", "false").lower() in ("true", "1", "yes")
        self.environment = os.getenv("ENVIRONMENT", "development")

        # Cloudflare R2 settings
        self.r2_account_id = os.getenv("R2_ACCOUNT_ID", "")
        self.r2_access_key_id = os.getenv("R2_ACCESS_KEY_ID", "")
        self.r2_secret_access_key = os.getenv("R2_SECRET_ACCESS_KEY", "")
        self.r2_bucket_name = os.getenv("R2_BUCKET_NAME", "")
        self.r2_public_url = os.getenv("R2_PUBLIC_URL", "")


settings = Settings()
